requirements.txt
alpaca-py
plotly
matplotlib
pandas
cufflinks
chart_studio

lumibot
numpy
pandas-datareader
quantstats
requests
yfinance
vectorbt